requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.34.125
botocore==1.34.125
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
diff-match-patch==20230430
Django==4.2.13
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-import-export==4.0.8
django-rest-swagger==2.2.0
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.1
prompt-toolkit==3.0.47
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
s3transfer==0.10.1
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
twisted==24.3.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2